[SWM-317] Add : google-analytics 4
