+++ src/app/layout.tsx
@@ -6,8 +6,9 @@ import { Inter } from 'next/font/google';
 import { Toaster } from 'react-hot-toast';
 import ResolutionCheck from '../components/ui/ResolutionCheck';
 import Footer from '../components/fnb/Footer';
-import ChannelTalk from '../components/ChannelTalk/ChannelTalk';
-import ChannelTalkManager from '../components/ChannelTalk/ChannelTalkManager';
+import ChannelTalk from '../components/tools/ChannelTalk/ChannelTalk';
+import ChannelTalkManager from '../components/tools/ChannelTalk/ChannelTalkManager';
+import GoogleAnalytics from '../components/tools/GoogleAnalytics/GoogleAnalytics';
 
 const inter = Inter({ subsets: ['latin'] });
 
@@ -25,6 +26,7 @@ export default function RootLayout({ children, modal }: Props) {
   return (
     <html lang='ko'>
       <body className={inter.className}>
+        <GoogleAnalytics />
         <ChannelTalk />
         <AuthSessionProvider>
           <QueryProvider>
